// src/helpers/types.ts
export enum SERVER_MESSAGE_TYPES {
    INIT = "init",
    PLAYER_LEAVE = "player_leave",

    // Parcels
    SELECT_PARCEL = "select_parcel",
    REMOVE_PARCEL = "remove_parcel",

    // Player
    PLAY_MODE_CHANGED = 'play_mode_changed',
    SELECTED_SCENE_ASSET = 'selected_scene_asset',
    EDIT_SCENE_ASSET = 'edit_scene_asset',
    EDIT_SCENE_ASSET_DONE = 'edit_scene_asset_done',
    EDIT_SCENE_ASSET_CANCEL = 'edit_scene_asset_cancel',
    PLACE_SELECTED_ASSET = 'place_asset',
    SELECT_CATALOG_ASSET = 'select_catalog_asset',
    PLAYER_CANCELED_CATALOG_ASSET = 'player_canceled_catalog',
    ASSET_OVER_SCENE_LIMIT = 'asset_over_scene_limit',
    SUBMIT_FEEDBACK = 'submit_feedback',
    PLAYER_SETTINGS = 'player_settings',
    FIRST_TIME = 'first_time',
    REMOVE_WEAPONS_BUILD_MODE = 'remove_weapons_build_mode',

    // Catalog and Assets
    CATALOG_UPDATED = 'catalog_updated',
    PLAYER_ASSET_UPLOADED = 'player_asset_uploaded',
    PLAYER_ASSET_CATALOG = 'player_asset_catalog',
    PLAYER_CATALOG_DEPLOYED = 'player_catalog_deployed',
    PLAYER_RECEIVED_MESSAGE = 'player_received_message',
    PLAYER_SCENES_CATALOG = 'player_scenes_catalog',
    PLAYER_JOINED_USER_WORLD = 'player_joined_private_world',
    PLAYER_EDIT_ASSET = 'player_edit_asset',
    UPDATE_ITEM_COMPONENT = "update_component",
    UPDATE_GRAB_Y_AXIS = 'update_grab_y_axix',
    PLAYER_ASSET_PENDING = 'player_asset_pending',
    UPDATE_ASSET_LOCKED = 'update_asset_locked',
    UPDATE_ASSET_BUILD_VIS = 'update_asset_build_visibility',
    DELETE_UGC_ASSET = 'delete_ugc_asset',
    ADD_WORLD_ASSETS = 'add_world_assets',
    DELETE_WORLD_ASSETS = 'delete_world_assets',
    CHUNK_SEND_ASSETS ='chunk_send_assets',

    // Scene
    SCENE_SAVE_NEW = "scene_save_new",
    SCENE_ADD_ITEM = 'scene_add_item',
    SCENE_DROPPED_GRABBED = 'scene_dropped_grabbed',
    SCENE_ADDED_NEW = "scene_added_new",
    SCENE_LOAD = 'scene_load',
    SCENE_UPDATE_ITEM = 'scene_update_item',
    SCENE_DELETE_ITEM = 'scene_delete_item',
    SCENE_ADD_BP = 'add_build_permissions',
    SCENE_DELETE_BP = 'delete_build_permissions',
    SCENE_DELETE = 'delete_scene',
    SCENE_UPDATE_PARCELS = 'scene_update_parcels',
    SCENE_SAVE_EDITS = 'scene_save_edits',
    SCENE_UPDATE_ENABLED = 'scene_update_enabled',
    SCENE_UPDATE_PRIVACY = 'scene_update_privacy',
    SCENE_DOWNLOAD = 'scene_download',
    SCENE_DEPLOY = 'scene_deploy',
    SCENE_DEPLOY_READY = 'scene_deploy_ready',
    SCENE_ADDED_SPAWN = "scene_added_spawn",
    SCENE_DELETE_SPAWN = "scene_delete_spawn",
    SCENE_CLEAR_ASSETS = 'scene_clear_assets',
    SCENE_DEPLOY_FINISHED = 'scene_deploy_finished',
    SCENE_ACTION = 'scene_action',
    SCENE_DELETE_GRABBED_ITEM = 'scene_delete_grabbed_item',
    SCENE_CREATE_QUEST = 'scene_create_quest',
    SCENE_DELETE_QUEST = 'scene_delete_quest',
    
    //World
    INIT_WORLD = "init_world",
    NEW_WORLD_CREATED = 'new_world_created',
    FORCE_DEPLOYMENT = 'force_deployment',
    SCENE_COUNT  = 'scene_count',
    ADDED_TUTORIAL = 'added_tutorial',
    REMOVED_TUTORIAL = 'removed_tutorial',
    UPDATED_TUTORIAL_CID = 'updated_tutorial_cid',
    WORLD_TRAVEL = 'world_travel',
    WORLD_ADD_BP = 'world_add_build_permissions',
    WORLD_DELETE_BP = 'world_delete_build_permissions',
    WORLD_ADD_BAN = 'world_add_ban',
    WORLD_DELETE_BAN = 'world_delete_ban',
    GET_MARKETPLACE = 'get_marketplace',
    START_GAME = 'start_game',
    END_GAME = 'end_game',
    FORCE_BACKUP = 'force_backup',
    EXPORT_WORLD = 'export_world',

    CUSTOM = "custom",//
    IWB_VERSION_UPDATE ='iwb_version_update',

    //AUDIUS SPECIFIC ACTIONS
    PLAY_AUDIUS_TRACK = 'play_audius_track',

    //REMOTE SERVER ACTIONS
    CLAIM_REWARD = 'claim_reward',
    VERIFY_ACCESS = 'verify_access',

    //GAMING
    CREATE_GAME_LOBBY = 'create_game_lobby',
    HIT_OBJECT = 'hit_object',
    SHOOT = 'shoot',
    LEADERBOARD_UPDATE = 'leaderboard_update',
    GAME_ACTION = 'game_action',

    //QUESTING
    GET_QUEST_DEFINITIONS = 'get_quest_definitions',
    QUEST_EDIT = 'edit_quest',
    QUEST_ACTION = 'quest_action',
    QUEST_PLAYER_DATA = 'quest_player_data',
    QUEST_STEP_DATA = 'quest_step_data',



    //conference//
    CONFERENCE_IMAGE_UPDATE = 'conference_image_update',
    CONFERENCE_VIDEO_UPDATE = 'conference_video_update',
    CONFERENCE_CLEAR = 'clear-conference',
    CONFERENCE_REFRESH = 'refresh-conference',
    CONFERENCE_ADD_RESERVATION = 'new-conference-reservation',
    CONFERENCE_CANCEL_RESERVATION = 'cancel-conference-reservation',
}